+++ data/parser.go
@@ -1,6 +1,8 @@
 package data
 
 import (
+	"time"
+
 	"github.com/tidepool-org/platform/log"
 	"github.com/tidepool-org/platform/service"
 )
@@ -19,6 +21,7 @@ type ObjectParser interface {
 	ParseFloat(key string) *float64
 	ParseString(key string) *string
 	ParseStringArray(key string) *[]string
+	ParseTime(key string, layout string) *time.Time
 	ParseObject(key string) *map[string]interface{}
 	ParseObjectArray(key string) *[]map[string]interface{}
 	ParseInterface(key string) *interface{}
@@ -44,6 +47,7 @@ type ArrayParser interface {
 	ParseFloat(index int) *float64
 	ParseString(index int) *string
 	ParseStringArray(index int) *[]string
+	ParseTime(index int, layout string) *time.Time
 	ParseObject(index int) *map[string]interface{}
 	ParseObjectArray(index int) *[]map[string]interface{}
 	ParseInterface(index int) *interface{}